Move `CallFrame` out of `Interpreter`.
